Rename user.UserGroupResolver to usergroup.Resolver
